backend/requirements.txt
# AgentLab Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WebSocket Support
python-socketio==5.10.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# AI/ML Libraries
cerebras-cloud-sdk==1.6.0
torch==2.2.0
transformers==4.35.2

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Logging and Monitoring
structlog==23.2.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Validation
email-validator==2.1.0

# Production ASGI Server
gunicorn==21.2.0

# Additional utilities
python-multipart==0.0.6
jinja2==3.1.2